+++ user.go
@@ -6,3 +6,10 @@ type User struct {
 }
 
 type Users []User
+
+type UserRole struct {
+	user_id int `json:"user_id"`
+	role_id int `json:"role_id"`
+}
+
+type UserRoles []UserRole
