--- user_roles.go
@@ -1,8 +1,0 @@
-package main
-
-type UserRole struct {
-	user_id int `json:"user_id"`
-	role_id int `json:"role_id"`
-}
-
-type UserRoles []UserRole
